requirements.txt
# DesktopClient Requirements
# Core classification engine (install editable from local package)
# pip install -e D:\Projects\PyCharmProjects\PipModules\Classification\PDFClassificationEngine

# Classification engine dependencies (also in PDFClassificationEngine)
PyMuPDF>=1.23.0
anthropic>=0.20.0
Pillow>=10.0.0

# Desktop tooling
python-dotenv>=1.0.0

# Build
pyinstaller>=6.0.0
